requirements.txt
alembic==1.18.4
babel==2.18.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
defusedxml==0.7.1
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Limiter==4.1.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.61.1
fpdf2==2.8.5
greenlet==3.3.1
gunicorn==23.0.0; platform_system != "Windows"
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
waitress==3.0.2
Werkzeug==3.1.6
pymysql
mysqlclient
cryptography
